fix: ise long for occurrence counts instead of int

+++ packages/graphql-api/src/resources/gbif/occurrence/occurrence.type.js
@@ -649,7 +649,7 @@ const typeDef = gql`
     maxLng: Int!
     minYear: Int
     maxYear: Int
-    total: Int!
+    total: Long!
     generated: DateTime!
   }
 `;
